content-agent-langgraph/requirements.txt
requests
beautifulsoup4
langchain
langchain-openai
python-dotenv
tweepy
openai
faiss-cpu
pytz